pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-skyfi"
version = "0.1.0"
description = "MCP server for SkyFi satellite imagery API"
readme = "README.md"
license = "MIT"
authors = [
    { name = "SkyFi Development Team", email = "dev@skyfi.com" },
]
requires-python = ">=3.10"
dependencies = [
    "mcp>=0.9.0",
    "httpx>=0.25.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "sse-starlette>=1.6.0",
    "redis>=5.0.0",
    "sqlalchemy>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "pytest-httpx>=0.30.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.scripts]
mcp-skyfi = "mcp_skyfi:main"

[tool.hatch.build]
packages = ["src/mcp_skyfi"]

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true